US975260: Refactor to remove HPE references in package names
https://internal.almoctane.com/ui/entity-navigation?p=131002/6001&entityType=work_item&id=975260
