+++ pom.xml
@@ -57,10 +57,10 @@
     </developers>
 
     <properties>
-        <caf.worker-framework.version>8.1.0-SNAPSHOT</caf.worker-framework.version>
+        <caf.worker-framework.version>9.0.0-1293</caf.worker-framework.version>
         <copyrightYear>2024</copyrightYear>
         <copyrightNotice>Copyright ${project.inceptionYear}-${copyrightYear} Open Text.</copyrightNotice>
-        <worker-batch-version>3.6.0-88</worker-batch-version>
+        <worker-batch-version>4.0.0-SNAPSHOT</worker-batch-version>
         <dockerHubOrganization>jobservice</dockerHubOrganization>
         <dockerJobServiceOrg>${dockerImagePrefix}${dockerHubOrganization}${dockerOrgSeperator}</dockerJobServiceOrg>
         <dockerProjectVersion>${dockerVersionSeperator}${project.version}</dockerProjectVersion>
@@ -80,22 +80,22 @@
             <dependency>
                 <groupId>com.github.workerframework.testing</groupId>
                 <artifactId>workerframework-testing-util</artifactId>
-                <version>3.1.0-SNAPSHOT</version>
+                <version>4.0.0-267</version>
             </dependency>
             <dependency>
                 <groupId>com.github.workerframework.testing</groupId>
                 <artifactId>workerframework-testing-integration</artifactId>
-                <version>3.1.0-SNAPSHOT</version>
+                <version>4.0.0-267</version>
             </dependency>
             <dependency>
-                <groupId>com.github.cafdataprocessing</groupId>
+                <groupId>com.github.cafdataprocessing.workers.document</groupId>
                 <artifactId>worker-document-shared</artifactId>
-                <version>2.0.0-5</version>
+                <version>7.0.0-1496</version>
             </dependency>
             <dependency>
                 <groupId>com.github.jobservice</groupId>
                 <artifactId>document-message-builder</artifactId>
-                <version>2.1.1-7</version>
+                <version>3.0.0-SNAPSHOT</version>
             </dependency>
             <dependency>
                 <groupId>com.github.jobservice</groupId>
