+++ release-notes-2.3.0.md
@@ -4,6 +4,7 @@ ${version-number}
 
 #### New Features
 - **SCMOD-9780**: Updated image to use Java 11
+- US975260: Refactor to remove HPE references in package names
 
 #### Known Issues
 - None
