+++ worker-globfilter-batch-plugin/src/main/java/com/github/jobservice/workers/batch/plugins/GlobFilterWorkerConstants.java
@@ -13,23 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hpe.caf.worker.batch.plugins;
+package com.github.jobservice.workers.batch.plugins;
 
-public final class GlobFilterDocument {
+public final class GlobFilterWorkerConstants {
 
-    private final String storageReference;
-    private final String fileName;
+    private GlobFilterWorkerConstants() {}
 
-    public GlobFilterDocument(final String storageReference, final String fileName){
-        this.storageReference = storageReference;
-        this.fileName = fileName;
-    }
+    public static final String DATASTORE_PARTIAL_REFERENCE_KEY = "datastorePartialReference";
 
-    public String getStorageReference(){
-        return this.storageReference;
-    }
-
-    public String getFileName() {
-        return this.fileName;
-    }
 }
