+++ worker-globfilter-batch-plugin/src/main/java/com/github/jobservice/workers/batch/plugins/GlobPattern.java
@@ -13,15 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hpe.caf.worker.batch.plugins;
+package com.github.jobservice.workers.batch.plugins;
 
-import com.hpe.caf.api.worker.DataStore;
-import com.hpe.caf.api.worker.DataStoreException;
-import com.hpe.caf.messagebuilder.TaskMessage;
-import com.hpe.caf.worker.batch.BatchDefinitionException;
-import com.hpe.caf.worker.batch.BatchWorkerPlugin;
-import com.hpe.caf.worker.batch.BatchWorkerServices;
-import com.hpe.caf.worker.batch.plugins.messagebuilder.GlobFilterWorkerTaskMessageBuilder;
+import com.github.jobservice.messagebuilder.TaskMessage;
+import com.github.jobservice.worker.batch.BatchDefinitionException;
+import com.github.jobservice.worker.batch.BatchWorkerPlugin;
+import com.github.jobservice.worker.batch.BatchWorkerServices;
+import com.github.jobservice.workers.batch.plugins.messagebuilder.GlobFilterWorkerTaskMessageBuilder;
+import com.github.workerframework.api.DataStore;
+import com.github.workerframework.api.DataStoreException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
