+++ worker-globfilter-batch-plugin/src/main/resources/META-INF/services/com.github.jobservice.workers.batch.BatchWorkerPlugin
@@ -1,0 +1,1 @@
+com.github.jobservice.workers.batch.plugins.GlobPattern
