+++ worker-globfilter-container/pom.xml
@@ -47,7 +47,7 @@
             <artifactId>worker-globfilter-batch-plugin</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.github.cafapi</groupId>
+            <groupId>com.github.cafapi.common</groupId>
             <artifactId>caf-api</artifactId>
         </dependency>
         <dependency>
@@ -65,7 +65,7 @@
             </exclusions>
         </dependency>
         <dependency>
-            <groupId>com.github.cafapi.config</groupId>
+            <groupId>com.github.cafapi.common</groupId>
             <artifactId>${caf.config.impl}</artifactId>
         </dependency>
         <dependency>
@@ -77,11 +77,11 @@
             <artifactId>${caf.worker-queue.impl}</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.github.cafapi.codec</groupId>
+            <groupId>com.github.cafapi.common</groupId>
             <artifactId>${caf.codec.impl}</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.github.cafapi.decoder</groupId>
+            <groupId>com.github.cafapi.common</groupId>
             <artifactId>decoder-js</artifactId>
         </dependency>
         <dependency>
